+++ src/rust/ffi/verilator_bridge.h
@@ -21,17 +21,17 @@ void vnpcsoc_set_reset(VNPCSoC* top, uint8_t val);
 
 uint8_t  vnpcsoc_get_probe_valid(const VNPCSoC* top);
 uint8_t  vnpcsoc_get_probe_is_mmio(const VNPCSoC* top);
-uint32_t vnpcsoc_get_probe_pc(const VNPCSoC* top);
-uint32_t vnpcsoc_get_probe_dnpc(const VNPCSoC* top);
+uint64_t vnpcsoc_get_probe_pc(const VNPCSoC* top);
+uint64_t vnpcsoc_get_probe_dnpc(const VNPCSoC* top);
 uint32_t vnpcsoc_get_probe_inst(const VNPCSoC* top);
-uint32_t vnpcsoc_get_probe_gpr(const VNPCSoC* top, int index);
-uint32_t vnpcsoc_get_probe_csr_mstatus(const VNPCSoC* top);
-uint32_t vnpcsoc_get_probe_csr_mtvec(const VNPCSoC* top);
-uint32_t vnpcsoc_get_probe_csr_mepc(const VNPCSoC* top);
-uint32_t vnpcsoc_get_probe_csr_mcause(const VNPCSoC* top);
-uint32_t vnpcsoc_get_probe_csr_mtval(const VNPCSoC* top);
-uint32_t vnpcsoc_get_probe_csr_mvendorid(const VNPCSoC* top);
-uint32_t vnpcsoc_get_probe_csr_marchid(const VNPCSoC* top);
+uint64_t vnpcsoc_get_probe_gpr(const VNPCSoC* top, int index);
+uint64_t vnpcsoc_get_probe_csr_mstatus(const VNPCSoC* top);
+uint64_t vnpcsoc_get_probe_csr_mtvec(const VNPCSoC* top);
+uint64_t vnpcsoc_get_probe_csr_mepc(const VNPCSoC* top);
+uint64_t vnpcsoc_get_probe_csr_mcause(const VNPCSoC* top);
+uint64_t vnpcsoc_get_probe_csr_mtval(const VNPCSoC* top);
+uint64_t vnpcsoc_get_probe_csr_mvendorid(const VNPCSoC* top);
+uint64_t vnpcsoc_get_probe_csr_marchid(const VNPCSoC* top);
 
 uint32_t vnpcsoc_get_probe_perf_commit_cnt(const VNPCSoC* top);
 uint32_t vnpcsoc_get_probe_perf_branch_cnt(const VNPCSoC* top);
